# openapi.yaml
openapi: 3.1.0
info:
  title: RescueBite API
  description: Production-grade RescueBite backend with restaurant operations, inventory
    expiry, reservations, and manual seller-confirmed payments.
  version: 1.1.0
paths:
...
  /restaurants/{restaurant_id}/batches:
    get:
      tags:
      - restaurants
      summary: Restaurant Batches
      operationId: restaurant_batches_restaurants__restaurant_id__batches_get
      parameters:
      - name: restaurant_id
        in: path
        required: true
        schema:
          type: integer
          title: Restaurant Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
...
  /batches:
    get:
      tags:
      - batches
      summary: List Batches
      operationId: list_batches_batches_get
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
    post:
      tags:
      - batches
      summary: Create Batch
      operationId: create_batch_batches_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /batches/{batchId}:
    get:
      tags:
      - batches
      summary: Get Batch
      operationId: get_batch_batches__batchId__get
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          title: Batchid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
    patch:
      tags:
      - batches
      summary: Update Batch
      operationId: update_batch_batches__batchId__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          title: Batchid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - batches
      summary: Delete Batch
      operationId: delete_batch_batches__batchId__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: integer
          title: Batchid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '400':
          description: Standard 400 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Standard 401 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Standard 403 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Standard 404 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Standard 409 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Standard 500 error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Allergen:
      type: string
      enum:
      - peanuts
      - tree_nuts
      - dairy
      - eggs
      - gluten
      - soy
      - fish
      - shellfish
      - sesame
      title: Allergen
...
    BatchCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
          default: ''
        category:
          type: string
          title: Category
        quantity_total:
          type: integer
          exclusiveMinimum: 0.0
          title: Quantity Total
        original_price_kzt:
          type: integer
          minimum: 0.0
          title: Original Price Kzt
        expires_at:
          type: string
          format: date-time
          title: Expires At
        pickup_start_at:
          type: string
          format: date-time
          title: Pickup Start At
        pickup_end_at:
          type: string
          format: date-time
          title: Pickup End At
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
        ingredients:
          items:
            $ref: '#/components/schemas/IngredientIn'
          type: array
          title: Ingredients
      type: object
      required:
      - title
      - category
      - quantity_total
      - original_price_kzt
      - expires_at
      - pickup_start_at
      - pickup_end_at
      - lat
      - lng
      - ingredients
      title: BatchCreate
    BatchList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BatchRead'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      title: BatchList
    BatchRead:
      properties:
        id:
          type: integer
          title: Id
        restaurant_id:
          type: integer
          title: Restaurant Id
        title:
          type: string
          title: Title
        category:
          type: string
          title: Category
        quantity_total:
          type: integer
          title: Quantity Total
        quantity_available:
          type: integer
          title: Quantity Available
        original_price_kzt:
          type: integer
          title: Original Price Kzt
        current_price_kzt:
          type: integer
          title: Current Price Kzt
        discount_percentage:
          type: integer
          title: Discount Percentage
        state:
          $ref: '#/components/schemas/BatchState'
        status:
          $ref: '#/components/schemas/BatchStatus'
        expires_at:
          type: string
          format: date-time
          title: Expires At
        expires_in_seconds:
          type: integer
          title: Expires In Seconds
        sale_label:
          type: string
          title: Sale Label
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
      type: object
      required:
      - id
      - restaurant_id
      - title
      - category
      - quantity_total
      - quantity_available
      - original_price_kzt
      - current_price_kzt
      - discount_percentage
      - state
      - status
      - expires_at
      - expires_in_seconds
      - sale_label
      - lat
      - lng
      title: BatchRead
    BatchState:
      type: string
      enum:
      - fresh
      - discounted
      - free
      - compost
      title: BatchState
    BatchStatus:
      type: string
      enum:
      - active
      - paused
      - withdrawn
      - deleted
      title: BatchStatus
    BatchUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        quantity_total:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Quantity Total
        original_price_kzt:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Original Price Kzt
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        pickup_start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Pickup Start At
        pickup_end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Pickup End At
        status:
          anyOf:
          - $ref: '#/components/schemas/BatchStatus'
          - type: 'null'
      type: object
      title: BatchUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngredientIn:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        allergen:
          anyOf:
          - $ref: '#/components/schemas/Allergen'
          - type: 'null'
      type: object
      required:
      - name
      title: IngredientIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      example:
        code: forbidden
        message: insufficient role
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT